Fix collision detection system
Fixes a bug when the I shape was rotated with 2 blank columns on
the left side that caused an overflow subtraction making the
game to panic in dev mode and the shape to dissapear in release
mode. Now negative column numbers should be correctly managed
with an isize variable and the playfield should also be able
to manage those indices correctly.

When fixing this bug, another related bug was detected: collision
detection caused out of bounds error when shapes next to the side
borders tried to rotate. When we tried to rotate I shape next to
the edges, some valid blocks ended out of bounds and the algorithm
tried to access those invalid playfield positions (col < 0,
col > 11, row > 22) instead of returning a collision.