Add Random Shape Generator
Adds a Random Shape Generator module to compute the number of
the next shape. Up to now there is only one Pseudo Random
Shape Generator.
Adds a ShapeFactory to Shape module to manage the creation of the
next shape based on the Generator . This required an important
refactoring of the Shape module to a more idiomatic approach,
where we use traits and enums implementing those traits to provide
different functionality depending on the Rotation System (used to
decide how to build the base shapes) and Type (used to decide how
to rotate those shapes).